Charades/backend/app/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Werkzeug==2.3.7